% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pcpr2}
\alias{plot.pcpr2}
\title{Plot PC-PR2 output}
\usage{
\method{plot}{pcpr2}(x, ...)
}
\arguments{
\item{x}{an object of class \code{pcpr2}.}

\item{...}{additional arguments passed to the
\code{\link{barplot}} function.}
}
\description{
\code{plot.pcpr2} plots partial R2 values.
}
\details{
This function plots a barchart of partial R2 values
created by the \code{\link{pcpr2}} function for easy interpretation.
The bars are annotated with values.
}
\examples{
output <- runPCPR2(transcripts, Z_metadata)
plot(output, col="red", main="Variability in transcriptomics data explained by covariates",
ylab="Rpartial2")
}
