% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPCPR2.R
\name{plotProp}
\alias{plotProp}
\title{Plot PC-PR2 output (to be deprecated)}
\usage{
plotProp(output, ...)
}
\arguments{
\item{...}{Other arguments passed to barplot().}

\item{Rpartial2}{Named vector of partial R2 values generated from runPCPR2().}
}
\description{
A wrapper for barplot() that plots PC-PR2 output. To shortly be replaced by an S3 class plot method.
}
\examples{
plotProp(output)
}
\seealso{
\code{\link{barplot}} which this function wraps.
}
