% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.pcpr2}
\alias{print.pcpr2}
\title{Print method for pcpr2}
\usage{
\method{print}{pcpr2}(x, ...)
}
\arguments{
\item{x}{an object of class \code{pcpr2}.}

\item{...}{additional arguments passed to the function.}
}
\value{
the input object is returned silently.
}
\description{
\code{print.pcpr2} Prints a mimimal output.
}
\details{
This function prints the model fit and Partial R2 values calculated
by the \code{\link{runPCPR2}} function.
}
\examples{
output <- runPCPR2(transcripts, Z_metadata)
print(output)
}
