% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.pcpr2}
\alias{summary.pcpr2}
\title{Summary method for pcpr2}
\usage{
\method{summary}{pcpr2}(x, ...)
}
\arguments{
\item{x}{an object of class \code{pcpr2}.}

\item{...}{additional arguments passed to the function.}
}
\value{
the input object is returned silently.
}
\description{
\code{print.pcpr2} Prints a detailed summary of the PC-PR2 object
}
\details{
This summary function prints a formatted summary the data supplied, the parameters supplied,
the model fit and the Partial R2 values calculated by the \code{\link{runPCPR2}} function.
}
\examples{
output <- runPCPR2(transcripts, Z_metadata)
summary(output)
}
