% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPCPR2.R
\name{plotProp}
\alias{plotProp}
\title{Plot PC-PR2 output (to be deprecated in favour of an S3 class plot method)}
\usage{
plotProp(Rpartial2, ...)
}
\arguments{
\item{Rpartial2}{Named vector of partial R2 values generated from runPCPR2().}

\item{...}{Other arguments passed to barplot().}
}
\description{
A wrapper for barplot() that plots PC-PR2 output.
}
\examples{
plotProp(output)
}
\seealso{
\code{\link{barplot}} which this function wraps.
}
