% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPCPR2.R
\name{runPCPR2}
\alias{runPCPR2}
\title{The Principal Component R-squared method (PC-PR2)}
\usage{
runPCPR2(X, Z, pct.threshold = 0.8)
}
\arguments{
\item{X}{A matrix of omics data. It is recommended that the data be appropriately scaled and log transformed if necessary.}

\item{Z}{The metadata variables, with the same number of observations as X, whose influence on the omics data (X-matrix) is to be assessed. Categorical variables should be coded as factors.}

\item{pct.threshold}{The proportion of variability desired to be explained. Defaults to 0.8.}
}
\value{
Returns a list with 5 elements:
\item{dimensions}{dimensions of the -omics matrix}
\item{types}{data types for each Z-variable}
\item{threshold}{the chosen variability to be explained}
\item{pcn}{the number of principal components used for the chosen threshold}
\item{pR2}{Rpartial2 values for each Z-variable and the overall R2}
}
\description{
Runs the PC-PR2 method on X, an omics matrix of intensities, and Z, the subject metadata to be assessed in the model.
}
\examples{
output <- runPCPR2(transcripts, Z_metadata)
output$pR2
}
\keyword{Rpartial2,}
\keyword{analysis,}
\keyword{component}
\keyword{metabolomics,}
\keyword{omics,}
\keyword{pca,}
\keyword{pcpr2,}
\keyword{principal}
\keyword{transcriptomics}
